+++ src/lang-xquery/test/uk/co/reecedunn/intellij/plugin/xquery/tests/lang/highlighter/XQueryColorSettingsPageTest.kt
@@ -89,12 +89,13 @@ class XQueryColorSettingsPageTest {
         assertThat(keys.contains(XQuerySyntaxHighlighterColors.XML_ATTRIBUTE_VALUE), `is`(true))
         assertThat(keys.contains(XQuerySyntaxHighlighterColors.XML_ENTITY_REFERENCE), `is`(true))
         assertThat(keys.contains(XQuerySyntaxHighlighterColors.XML_ESCAPED_CHARACTER), `is`(true))
+        assertThat(keys.contains(XQuerySyntaxHighlighterColors.XML_PI_TAG), `is`(true))
         assertThat(keys.contains(XQuerySyntaxHighlighterColors.XML_TAG), `is`(true))
         assertThat(keys.contains(XQuerySyntaxHighlighterColors.XML_TAG_NAME), `is`(true))
         assertThat(keys.contains(XQuerySyntaxHighlighterColors.XQDOC_MARKUP), `is`(true))
         assertThat(keys.contains(XQuerySyntaxHighlighterColors.XQDOC_TAG), `is`(true))
         assertThat(keys.contains(XQuerySyntaxHighlighterColors.XQDOC_TAG_VALUE), `is`(true))
-        assertThat(keys.size, `is`(19)) // No other matching highlight colours
+        assertThat(keys.size, `is`(20)) // No other matching highlight colours
     }
 
     @Test